requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==2.4.0
django-localflavor==3.1
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.3.2
python-stdnum==1.16
pytz==2021.1
sqlparse==0.4.2
